requirements.txt
setuptools~=67.6.0
requests~=2.28.2
untangle~=1.2.1
python-dateutil~=2.8.2